requirements.txt
email-validator
flask-migrate
flask-wtf
Werkzeug
Flask
alembic
SQLAlchemy
pillow
WTForms
pymorphy2
